refactor: use containerbase not buildpack terminology
