+++ lib/modules/manager/flux/readme.md
@@ -19,7 +19,7 @@ Namespaces will not be inferred from the context (e.g. from the parent `Kustomiz
 Updating system manifests requires that either:
 
 1. The `flux` tool is pre-installed, or
-2. You run a Docker image based on [containerbase/buildpack](https://github.com/containerbase/buildpack), such as the official Renovate images, and have `binarySource=install` configured
+2. You run a Docker image based on [containerbase](https://github.com/containerbase), such as the official Renovate images, and have `binarySource=install` configured
 
 By default, the `flux` manager will only match `flux-system/gotk-components.yaml` (i.e. system manifest) files.
 This is because there is no commonly accepted file/directory naming convention for Flux manifests and we don't want to check every single `*.yaml` file in repositories just in case some of them have Flux definitions.
